requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
altgraph==0.17.3
    # via pyinstaller
appdirs==1.4.4
    # via
    #   -r requirements/base.txt
    #   tutor
astroid==2.15.6
    # via pylint
asttokens==2.2.1
    # via
    #   -r requirements/base.txt
    #   transifex-python
bcrypt==4.0.1
    # via -r requirements/base.txt
black==23.7.0
    # via -r requirements/dev.in
bleach==6.0.0
    # via readme-renderer
cachetools==5.3.1
    # via
    #   -r requirements/base.txt
    #   google-auth
certifi==2023.7.22
    # via
    #   -r requirements/base.txt
    #   kubernetes
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.6
    # via
    #   -r requirements/base.txt
    #   black
    #   transifex-python
    #   tutor
cryptography==41.0.2
    # via secretstorage
dill==0.3.7
    # via pylint
docutils==0.20.1
    # via readme-renderer
future==0.18.3
    # via
    #   -r requirements/base.txt
    #   pyseeyou
google-auth==2.22.0
    # via
    #   -r requirements/base.txt
    #   kubernetes
idna==3.4
    # via
    #   -r requirements/base.txt
    #   requests
importlib-metadata==6.8.0
    # via
    #   keyring
    #   twine
importlib-resources==6.0.0
    # via keyring
isort==5.12.0
    # via pylint
jaraco-classes==3.3.0
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jinja2==3.1.2
    # via
    #   -r requirements/base.txt
    #   tutor
keyring==24.2.0
    # via twine
kubernetes==27.2.0
    # via
    #   -r requirements/base.txt
    #   tutor
lazy-object-proxy==1.9.0
    # via astroid
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via
    #   -r requirements/base.txt
    #   jinja2
mccabe==0.7.0
    # via pylint
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.0.0
    # via jaraco-classes
mypy==1.4.1
    # via
    #   -r requirements/base.txt
    #   tutor
mypy-extensions==1.0.0
    # via
    #   -r requirements/base.txt
    #   black
    #   mypy
oauthlib==3.2.2
    # via
    #   -r requirements/base.txt
    #   kubernetes
    #   requests-oauthlib
packaging==23.1
    # via black
parsimonious==0.10.0
    # via
    #   -r requirements/base.txt
    #   pyseeyou
pathspec==0.11.1
    # via black
pkginfo==1.9.6
    # via twine
platformdirs==3.9.1
    # via
    #   black
    #   pylint
pyasn1==0.5.0
    # via
    #   -r requirements/base.txt
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via
    #   -r requirements/base.txt
    #   google-auth
pycparser==2.21
    # via cffi
pycryptodome==3.18.0
    # via
    #   -r requirements/base.txt
    #   tutor
pygments==2.15.1
    # via
    #   readme-renderer
    #   rich
pyinstaller==5.13.0
    # via -r requirements/dev.in
pyinstaller-hooks-contrib==2023.6
    # via pyinstaller
pylint==2.17.5
    # via -r requirements/dev.in
pyseeyou==1.0.2
    # via
    #   -r requirements/base.txt
    #   transifex-python
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   kubernetes
pytz==2023.3
    # via
    #   -r requirements/base.txt
    #   transifex-python
pyyaml==6.0.1
    # via
    #   -r requirements/base.txt
    #   kubernetes
    #   tutor
readme-renderer==40.0
    # via twine
regex==2023.6.3
    # via
    #   -r requirements/base.txt
    #   parsimonious
requests==2.31.0
    # via
    #   -r requirements/base.txt
    #   kubernetes
    #   requests-oauthlib
    #   requests-toolbelt
    #   transifex-python
    #   twine
requests-oauthlib==1.3.1
    # via
    #   -r requirements/base.txt
    #   kubernetes
requests-toolbelt==1.0.0
    # via twine
rfc3986==2.0.0
    # via twine
rich==13.4.2
    # via twine
rsa==4.9
    # via
    #   -r requirements/base.txt
    #   google-auth
secretstorage==3.3.3
    # via keyring
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   asttokens
    #   bleach
    #   google-auth
    #   kubernetes
    #   python-dateutil
tomli==2.0.1
    # via
    #   -r requirements/base.txt
    #   black
    #   mypy
    #   pylint
tomlkit==0.12.1
    # via pylint
toolz==0.12.0
    # via
    #   -r requirements/base.txt
    #   pyseeyou
transifex-python==3.4.0
    # via -r requirements/base.txt
tutor==16.0.2
    # via -r requirements/base.txt
twine==4.0.2
    # via -r requirements/dev.in
typing-extensions==4.7.1
    # via
    #   -r requirements/base.txt
    #   astroid
    #   black
    #   mypy
    #   pylint
    #   rich
    #   tutor
urllib3==1.26.16
    # via
    #   -r requirements/base.txt
    #   google-auth
    #   kubernetes
    #   requests
    #   twine
webencodings==0.5.1
    # via bleach
websocket-client==1.6.1
    # via
    #   -r requirements/base.txt
    #   kubernetes
wrapt==1.15.0
    # via astroid
zipp==3.16.2
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
